requirements.txt
python_version == "3.11"
langchain==0.1.16
langchain-openai==0.1.3
langchain_community==0.0.21
openai==1.3.7
flask==2.3.3
python-dotenv==1.0.0
faiss-cpu==1.7.4
pypdf==3.17.4
pydantic==2.5.2
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.35.2